+++ go.mod
@@ -1,0 +1,27 @@
+module github.com/FreekingDean/docker-machine-driver-proxmox-ve
+
+go 1.21.4
+
+require (
+	github.com/FreekingDean/proxmox-api-go v0.2.0
+	github.com/docker/machine v0.16.2
+	github.com/labstack/gommon v0.4.2
+	github.com/lnxbil/docker-machine-driver-proxmox-ve v0.0.0-20200729213752-424d127ead78
+	github.com/mosolovsa/go_cat_sshfilerw v0.0.0-20170307095915-3f6912dbd9e4
+	golang.org/x/crypto v0.18.0
+	gopkg.in/resty.v1 v1.12.0
+)
+
+require (
+	github.com/Azure/go-ansiterm v0.0.0-20210617225240-d185dfc1b5a1 // indirect
+	github.com/docker/docker v20.10.27+incompatible // indirect
+	github.com/google/go-querystring v1.1.0 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/moby/term v0.5.0 // indirect
+	github.com/valyala/bytebufferpool v1.0.0 // indirect
+	github.com/valyala/fasttemplate v1.2.2 // indirect
+	golang.org/x/net v0.10.0 // indirect
+	golang.org/x/sys v0.16.0 // indirect
+	golang.org/x/term v0.16.0 // indirect
+)
